refactor(challenge): my-challenge 루트 orphan ParsedCommentBox를 ui/로
4개 하위기능(bonus-mission·ot-mission·mission·menu)이 공유하는 표현 컴포넌트가
my-challenge 루트에 떠 있어 my-challenge/ui/ 신설 후 이동. 루트 정리.
tsc 0 / lint 통과. 동작 변경 없음.
